Create BulkPasswordChange.ps1
Quick script that alters the passwords for a bulk amount of users and sets it to a new specified one. Useful for quick password changing when worrying about numerous users and wanting to change all passwords to the same one, temporarily. Has the ability of excluding the changes for specific users.